broken table down into components
