+++ frontend/src/components/Table/Pot/Pot.tsx
@@ -1,0 +1,9 @@
+type PotProps = {};
+
+export default function Pot({}: PotProps) {
+  return (
+    <div id="pot" data-chip-count="0">
+      Pot
+    </div>
+  );
+}
